move regressor.py from algorithms/operon to experiment/methods/operon
